requirements.txt
Flask==2.3.2
transformers==4.31.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.4
tokenizers-0.13.3